rename text-field class names
