+++ vaadin-text-field-flow-parent/vaadin-text-field-flow-integration-tests/src/main/java/com/vaadin/flow/component/textfield/tests/validation/NumberFieldBasicValidationPage.java
@@ -16,48 +16,22 @@
 package com.vaadin.flow.component.textfield.tests.validation;
 
 import com.vaadin.flow.component.textfield.NumberField;
-import com.vaadin.flow.data.binder.Binder;
 import com.vaadin.flow.router.Route;
 import com.vaadin.tests.validation.AbstractValidationPage;
 
-@Route("vaadin-number-field/validation/binder")
-public class NumberFieldValidationBinderPage
+@Route("vaadin-number-field/validation/basic")
+public class NumberFieldBasicValidationPage
         extends AbstractValidationPage<NumberField> {
+    public static final String REQUIRED_BUTTON = "required-button";
     public static final String STEP_INPUT = "step-input";
     public static final String MIN_INPUT = "min-input";
     public static final String MAX_INPUT = "max-input";
-    public static final String EXPECTED_VALUE_INPUT = "expected-value-input";
 
-    public static final String REQUIRED_ERROR_MESSAGE = "The field is required";
-    public static final String UNEXPECTED_VALUE_ERROR_MESSAGE = "The field doesn't match the expected value";
-
-    public static class Bean {
-        private Double property;
-
-        public Double getProperty() {
-            return property;
-        }
-
-        public void setProperty(Double property) {
-            this.property = property;
-        }
-    }
-
-    protected Binder<Bean> binder;
-
-    private Double expectedValue;
-
-    public NumberFieldValidationBinderPage() {
+    public NumberFieldBasicValidationPage() {
         super();
 
-        binder = new Binder<>(Bean.class);
-        binder.forField(testField).asRequired(REQUIRED_ERROR_MESSAGE)
-                .withValidator(value -> value.equals(expectedValue),
-                        UNEXPECTED_VALUE_ERROR_MESSAGE)
-                .bind("property");
-
-        add(createInput(EXPECTED_VALUE_INPUT, "Set expected value", event -> {
-            expectedValue = Double.parseDouble(event.getValue());
+        add(createButton(REQUIRED_BUTTON, "Enable required", event -> {
+            testField.setRequiredIndicatorVisible(true);
         }));
 
         add(createInput(STEP_INPUT, "Set step", event -> {
