+++ vaadin-text-field-flow-parent/vaadin-text-field-flow-integration-tests/src/main/java/com/vaadin/flow/component/textfield/tests/validation/NumberFieldBinderValidationPage.java
@@ -15,39 +15,39 @@
  */
 package com.vaadin.flow.component.textfield.tests.validation;
 
-import com.vaadin.flow.component.textfield.EmailField;
+import com.vaadin.flow.component.textfield.NumberField;
 import com.vaadin.flow.data.binder.Binder;
 import com.vaadin.flow.router.Route;
 import com.vaadin.tests.validation.AbstractValidationPage;
 
-@Route("vaadin-email-field/validation/binder")
-public class EmailFieldValidationBinderPage
-        extends AbstractValidationPage<EmailField> {
-    public static final String PATTERN_INPUT = "pattern-input";
-    public static final String MIN_LENGTH_INPUT = "min-length-input";
-    public static final String MAX_LENGTH_INPUT = "max-length-input";
+@Route("vaadin-number-field/validation/binder")
+public class NumberFieldBinderValidationPage
+        extends AbstractValidationPage<NumberField> {
+    public static final String STEP_INPUT = "step-input";
+    public static final String MIN_INPUT = "min-input";
+    public static final String MAX_INPUT = "max-input";
     public static final String EXPECTED_VALUE_INPUT = "expected-value-input";
 
     public static final String REQUIRED_ERROR_MESSAGE = "The field is required";
     public static final String UNEXPECTED_VALUE_ERROR_MESSAGE = "The field doesn't match the expected value";
 
     public static class Bean {
-        private String property;
+        private Double property;
 
-        public String getProperty() {
+        public Double getProperty() {
             return property;
         }
 
-        public void setProperty(String property) {
+        public void setProperty(Double property) {
             this.property = property;
         }
     }
 
     protected Binder<Bean> binder;
 
-    private String expectedValue;
+    private Double expectedValue;
 
-    public EmailFieldValidationBinderPage() {
+    public NumberFieldBinderValidationPage() {
         super();
 
         binder = new Binder<>(Bean.class);
@@ -57,25 +57,26 @@ public class EmailFieldValidationBinderPage
                 .bind("property");
 
         add(createInput(EXPECTED_VALUE_INPUT, "Set expected value", event -> {
-            expectedValue = event.getValue();
+            expectedValue = Double.parseDouble(event.getValue());
         }));
 
-        add(createInput(PATTERN_INPUT, "Set pattern", event -> {
-            testField.setPattern(event.getValue());
+        add(createInput(STEP_INPUT, "Set step", event -> {
+            double value = Double.parseDouble(event.getValue());
+            testField.setStep(value);
         }));
 
-        add(createInput(MIN_LENGTH_INPUT, "Set min length", event -> {
-            int value = Integer.parseInt(event.getValue());
-            testField.setMinLength(value);
+        add(createInput(MIN_INPUT, "Set min", event -> {
+            double value = Double.parseDouble(event.getValue());
+            testField.setMin(value);
         }));
 
-        add(createInput(MAX_LENGTH_INPUT, "Set max length", event -> {
-            int value = Integer.parseInt(event.getValue());
-            testField.setMaxLength(value);
+        add(createInput(MAX_INPUT, "Set max", event -> {
+            double value = Double.parseDouble(event.getValue());
+            testField.setMax(value);
         }));
     }
 
-    protected EmailField createTestField() {
-        return new EmailField();
+    protected NumberField createTestField() {
+        return new NumberField();
     }
 }
