+++ vaadin-text-field-flow-parent/vaadin-text-field-flow-integration-tests/src/main/java/com/vaadin/flow/component/textfield/tests/validation/TextAreaBasicValidationPage.java
@@ -15,42 +15,41 @@
  */
 package com.vaadin.flow.component.textfield.tests.validation;
 
-import com.vaadin.flow.component.textfield.NumberField;
+import com.vaadin.flow.component.textfield.TextArea;
 import com.vaadin.flow.router.Route;
 import com.vaadin.tests.validation.AbstractValidationPage;
 
-@Route("vaadin-number-field/validation/basic")
-public class NumberFieldValidationBasicPage
-        extends AbstractValidationPage<NumberField> {
+@Route("vaadin-text-area/validation/basic")
+public class TextAreaBasicValidationPage
+        extends AbstractValidationPage<TextArea> {
     public static final String REQUIRED_BUTTON = "required-button";
-    public static final String STEP_INPUT = "step-input";
-    public static final String MIN_INPUT = "min-input";
-    public static final String MAX_INPUT = "max-input";
+    public static final String PATTERN_INPUT = "pattern-input";
+    public static final String MIN_LENGTH_INPUT = "min-length-input";
+    public static final String MAX_LENGTH_INPUT = "max-length-input";
 
-    public NumberFieldValidationBasicPage() {
+    public TextAreaBasicValidationPage() {
         super();
 
         add(createButton(REQUIRED_BUTTON, "Enable required", event -> {
             testField.setRequiredIndicatorVisible(true);
         }));
 
-        add(createInput(STEP_INPUT, "Set step", event -> {
-            double value = Double.parseDouble(event.getValue());
-            testField.setStep(value);
+        add(createInput(PATTERN_INPUT, "Set pattern", event -> {
+            testField.setPattern(event.getValue());
         }));
 
-        add(createInput(MIN_INPUT, "Set min", event -> {
-            double value = Double.parseDouble(event.getValue());
-            testField.setMin(value);
+        add(createInput(MIN_LENGTH_INPUT, "Set min length", event -> {
+            int value = Integer.parseInt(event.getValue());
+            testField.setMinLength(value);
         }));
 
-        add(createInput(MAX_INPUT, "Set max", event -> {
-            double value = Double.parseDouble(event.getValue());
-            testField.setMax(value);
+        add(createInput(MAX_LENGTH_INPUT, "Set max length", event -> {
+            int value = Integer.parseInt(event.getValue());
+            testField.setMaxLength(value);
         }));
     }
 
-    protected NumberField createTestField() {
-        return new NumberField();
+    protected TextArea createTestField() {
+        return new TextArea();
     }
 }
