+++ vaadin-text-field-flow-parent/vaadin-text-field-flow-integration-tests/src/main/java/com/vaadin/flow/component/textfield/tests/validation/TextFieldBinderValidationPage.java
@@ -15,23 +15,49 @@
  */
 package com.vaadin.flow.component.textfield.tests.validation;
 
-import com.vaadin.flow.component.textfield.PasswordField;
+import com.vaadin.flow.component.textfield.TextField;
+import com.vaadin.flow.data.binder.Binder;
 import com.vaadin.flow.router.Route;
 import com.vaadin.tests.validation.AbstractValidationPage;
 
-@Route("vaadin-password-field/validation/basic")
-public class PasswordFieldValidationBasicPage
-        extends AbstractValidationPage<PasswordField> {
-    public static final String REQUIRED_BUTTON = "required-button";
+@Route("vaadin-text-field/validation/binder")
+public class TextFieldBinderValidationPage
+        extends AbstractValidationPage<TextField> {
     public static final String PATTERN_INPUT = "pattern-input";
     public static final String MIN_LENGTH_INPUT = "min-length-input";
     public static final String MAX_LENGTH_INPUT = "max-length-input";
+    public static final String EXPECTED_VALUE_INPUT = "expected-value-input";
 
-    public PasswordFieldValidationBasicPage() {
+    public static final String REQUIRED_ERROR_MESSAGE = "The field is required";
+    public static final String UNEXPECTED_VALUE_ERROR_MESSAGE = "The field doesn't match the expected value";
+
+    public static class Bean {
+        private String property;
+
+        public String getProperty() {
+            return property;
+        }
+
+        public void setProperty(String property) {
+            this.property = property;
+        }
+    }
+
+    protected Binder<Bean> binder;
+
+    private String expectedValue;
+
+    public TextFieldBinderValidationPage() {
         super();
 
-        add(createButton(REQUIRED_BUTTON, "Enable required", event -> {
-            testField.setRequiredIndicatorVisible(true);
+        binder = new Binder<>(Bean.class);
+        binder.forField(testField).asRequired(REQUIRED_ERROR_MESSAGE)
+                .withValidator(value -> value.equals(expectedValue),
+                        UNEXPECTED_VALUE_ERROR_MESSAGE)
+                .bind("property");
+
+        add(createInput(EXPECTED_VALUE_INPUT, "Set expected value", event -> {
+            expectedValue = event.getValue();
         }));
 
         add(createInput(PATTERN_INPUT, "Set pattern", event -> {
@@ -49,7 +75,7 @@ public class PasswordFieldValidationBasicPage
         }));
     }
 
-    protected PasswordField createTestField() {
-        return new PasswordField();
+    protected TextField createTestField() {
+        return new TextField();
     }
 }
