--- vaadin-text-field-flow-parent/vaadin-text-field-flow-integration-tests/src/test/java/com/vaadin/flow/component/textfield/tests/validation/EmailFieldValidationBinderIT.java
@@ -1,156 +1,0 @@
-/*
- * Copyright 2000-2022 Vaadin Ltd.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License. You may obtain a copy of
- * the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under
- * the License.
- */
-package com.vaadin.flow.component.textfield.tests.validation;
-
-import org.junit.Test;
-import org.openqa.selenium.Keys;
-
-import com.vaadin.flow.component.textfield.testbench.EmailFieldElement;
-import com.vaadin.flow.testutil.TestPath;
-import com.vaadin.tests.validation.AbstractValidationIT;
-
-import static com.vaadin.flow.component.textfield.tests.validation.EmailFieldValidationBinderPage.PATTERN_INPUT;
-import static com.vaadin.flow.component.textfield.tests.validation.EmailFieldValidationBinderPage.MIN_LENGTH_INPUT;
-import static com.vaadin.flow.component.textfield.tests.validation.EmailFieldValidationBinderPage.MAX_LENGTH_INPUT;
-import static com.vaadin.flow.component.textfield.tests.validation.EmailFieldValidationBinderPage.EXPECTED_VALUE_INPUT;
-import static com.vaadin.flow.component.textfield.tests.validation.EmailFieldValidationBinderPage.REQUIRED_ERROR_MESSAGE;
-import static com.vaadin.flow.component.textfield.tests.validation.EmailFieldValidationBinderPage.UNEXPECTED_VALUE_ERROR_MESSAGE;
-
-@TestPath("vaadin-email-field/validation/binder")
-public class EmailFieldValidationBinderIT
-        extends AbstractValidationIT<EmailFieldElement> {
-    @Test
-    public void fieldIsInitiallyValid() {
-        assertClientValid();
-        assertServerValid();
-        assertErrorMessage(null);
-    }
-
-    @Test
-    public void required_triggerInputBlur_assertValidity() {
-        testField.sendKeys(Keys.TAB);
-        assertServerInvalid();
-        assertClientInvalid();
-        assertErrorMessage(REQUIRED_ERROR_MESSAGE);
-    }
-
-    @Test
-    public void required_changeInputValue_assertValidity() {
-        $("input").id(EXPECTED_VALUE_INPUT).sendKeys("john@vaadin.com",
-                Keys.ENTER);
-
-        testField.setValue("john@vaadin.com");
-        assertServerValid();
-        assertClientValid();
-
-        testField.setValue("");
-        assertServerInvalid();
-        assertClientInvalid();
-        assertErrorMessage(REQUIRED_ERROR_MESSAGE);
-    }
-
-    @Test
-    public void minLength_changeInputValue_assertValidity() {
-        $("input").id(MIN_LENGTH_INPUT).sendKeys("13", Keys.ENTER);
-        $("input").id(EXPECTED_VALUE_INPUT).sendKeys("aaa@vaadin.com",
-                Keys.ENTER);
-
-        // Constraint validation fails:
-        testField.setValue("a@vaadin.com");
-        assertClientInvalid();
-        assertServerInvalid();
-        assertErrorMessage("");
-
-        // Binder validation fails:
-        testField.setValue("aa@vaadin.com");
-        assertClientInvalid();
-        assertServerInvalid();
-        assertErrorMessage(UNEXPECTED_VALUE_ERROR_MESSAGE);
-
-        // Both validations pass:
-        testField.setValue("aaa@vaadin.com");
-        assertClientValid();
-        assertServerValid();
-    }
-
-    @Test
-    public void maxLength_changeInputValue_assertValidity() {
-        $("input").id(MAX_LENGTH_INPUT).sendKeys("13", Keys.ENTER);
-        $("input").id(EXPECTED_VALUE_INPUT).sendKeys("a@vaadin.com",
-                Keys.ENTER);
-
-        // Constraint validation fails:
-        testField.setValue("aaa@vaadin.com");
-        assertClientInvalid();
-        assertServerInvalid();
-        assertErrorMessage("");
-
-        // Binder validation fails:
-        testField.setValue("aa@vaadin.com");
-        assertClientInvalid();
-        assertServerInvalid();
-        assertErrorMessage(UNEXPECTED_VALUE_ERROR_MESSAGE);
-
-        // Both validations pass:
-        testField.setValue("a@vaadin.com");
-        assertClientValid();
-        assertServerValid();
-    }
-
-    @Test
-    public void defaultPattern_changeInputValue_assertValidity() {
-        $("input").id(EXPECTED_VALUE_INPUT).sendKeys("john@vaadin.com",
-                Keys.ENTER);
-
-        testField.setValue("arbitrary string");
-        assertClientInvalid();
-        assertServerInvalid();
-        assertErrorMessage("");
-
-        testField.setValue("john@vaadin.com");
-        assertClientValid();
-        assertServerValid();
-    }
-
-    @Test
-    public void pattern_changeInputValue_assertValidity() {
-        $("input").id(PATTERN_INPUT).sendKeys("^[^\\d]+@vaadin.com$",
-                Keys.ENTER);
-        $("input").id(EXPECTED_VALUE_INPUT).sendKeys("john@vaadin.com",
-                Keys.ENTER);
-
-        // Constraint validation fails:
-        testField.setValue("2222@vaadin.com");
-        assertClientInvalid();
-        assertServerInvalid();
-        assertErrorMessage("");
-
-        // Binder validation fails:
-        testField.setValue("oliver@vaadin.com");
-        assertClientInvalid();
-        assertServerInvalid();
-        assertErrorMessage(UNEXPECTED_VALUE_ERROR_MESSAGE);
-
-        // Both validations pass:
-        testField.setValue("john@vaadin.com");
-        assertClientValid();
-        assertServerValid();
-    }
-
-    protected EmailFieldElement getTestField() {
-        return $(EmailFieldElement.class).first();
-    }
-}
