+++ vaadin-text-field-flow-parent/vaadin-text-field-flow-integration-tests/src/test/java/com/vaadin/flow/component/textfield/tests/validation/NumberFieldBinderValidationIT.java
@@ -1,0 +1,155 @@
+/*
+ * Copyright 2000-2022 Vaadin Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
+package com.vaadin.flow.component.textfield.tests.validation;
+
+import org.junit.Test;
+import org.openqa.selenium.Keys;
+
+import com.vaadin.flow.component.textfield.testbench.NumberFieldElement;
+import com.vaadin.flow.testutil.TestPath;
+import com.vaadin.tests.validation.AbstractValidationIT;
+
+import static com.vaadin.flow.component.textfield.tests.validation.NumberFieldBinderValidationPage.STEP_INPUT;
+import static com.vaadin.flow.component.textfield.tests.validation.NumberFieldBinderValidationPage.MIN_INPUT;
+import static com.vaadin.flow.component.textfield.tests.validation.NumberFieldBinderValidationPage.MAX_INPUT;
+import static com.vaadin.flow.component.textfield.tests.validation.NumberFieldBinderValidationPage.EXPECTED_VALUE_INPUT;
+import static com.vaadin.flow.component.textfield.tests.validation.NumberFieldBinderValidationPage.REQUIRED_ERROR_MESSAGE;
+import static com.vaadin.flow.component.textfield.tests.validation.NumberFieldBinderValidationPage.UNEXPECTED_VALUE_ERROR_MESSAGE;
+
+@TestPath("vaadin-number-field/validation/binder")
+public class NumberFieldBinderValidationIT
+        extends AbstractValidationIT<NumberFieldElement> {
+    @Test
+    public void fieldIsInitiallyValid() {
+        assertClientValid();
+        assertServerValid();
+        assertErrorMessage(null);
+    }
+
+    @Test
+    public void required_triggerInputBlur_assertValidity() {
+        testField.sendKeys(Keys.TAB);
+        assertServerInvalid();
+        assertClientInvalid();
+        assertErrorMessage(REQUIRED_ERROR_MESSAGE);
+    }
+
+    @Test
+    public void required_changeInputValue_assertValidity() {
+        $("input").id(EXPECTED_VALUE_INPUT).sendKeys("1234", Keys.ENTER);
+
+        testField.setValue("1234");
+        assertServerValid();
+        assertClientValid();
+
+        testField.setValue("");
+        assertServerInvalid();
+        assertClientInvalid();
+        assertErrorMessage(REQUIRED_ERROR_MESSAGE);
+    }
+
+    @Test
+    public void min_changeInputValue_assertValidity() {
+        $("input").id(MIN_INPUT).sendKeys("2", Keys.ENTER);
+        $("input").id(EXPECTED_VALUE_INPUT).sendKeys("2.2", Keys.ENTER);
+
+        // Constraint validation fails:
+        testField.setValue("1.8");
+        assertClientInvalid();
+        assertServerInvalid();
+        assertErrorMessage("");
+
+        // Binder validation fails:
+        testField.setValue("2");
+        assertClientInvalid();
+        assertServerInvalid();
+        assertErrorMessage(UNEXPECTED_VALUE_ERROR_MESSAGE);
+
+        // Both validations pass:
+        testField.setValue("2.2");
+        assertClientValid();
+        assertServerValid();
+    }
+
+    @Test
+    public void max_changeInputValue_assertValidity() {
+        $("input").id(MAX_INPUT).sendKeys("2", Keys.ENTER);
+        $("input").id(EXPECTED_VALUE_INPUT).sendKeys("1.8", Keys.ENTER);
+
+        // Constraint validation fails:
+        testField.setValue("2.2");
+        assertClientInvalid();
+        assertServerInvalid();
+        assertErrorMessage("");
+
+        // Binder validation fails:
+        testField.setValue("2");
+        assertClientInvalid();
+        assertServerInvalid();
+        assertErrorMessage(UNEXPECTED_VALUE_ERROR_MESSAGE);
+
+        // Both validations pass:
+        testField.setValue("1.8");
+        assertClientValid();
+        assertServerValid();
+    }
+
+    @Test
+    public void step_changeInputValue_assertValidity() {
+        $("input").id(STEP_INPUT).sendKeys("1.5", Keys.ENTER);
+        $("input").id(EXPECTED_VALUE_INPUT).sendKeys("3", Keys.ENTER);
+
+        // Constraint validation fails:
+        testField.setValue("1");
+        assertClientInvalid();
+        assertServerInvalid();
+        assertErrorMessage("");
+
+        // Binder validation fails:
+        testField.setValue("1.5");
+        assertClientInvalid();
+        assertServerInvalid();
+        assertErrorMessage(UNEXPECTED_VALUE_ERROR_MESSAGE);
+
+        // Both validations pass:
+        testField.setValue("3");
+        assertClientValid();
+        assertServerValid();
+    }
+
+    @Test
+    public void badInput_changeInputValue_assertValidity() {
+        $("input").id(EXPECTED_VALUE_INPUT).sendKeys("2", Keys.ENTER);
+
+        testField.sendKeys("--2", Keys.TAB);
+        assertServerInvalid();
+        assertClientInvalid();
+        assertErrorMessage("");
+
+        testField.setValue("2");
+        assertServerValid();
+        assertClientValid();
+
+        testField.sendKeys("--2", Keys.TAB);
+        assertServerInvalid();
+        assertClientInvalid();
+        assertErrorMessage("");
+    }
+
+    protected NumberFieldElement getTestField() {
+        return $(NumberFieldElement.class).first();
+    }
+}
